+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Alternatively view it [online on GitHub](https://github.com/cumulusmx/CumulusMX/
 - Two new web tags for the Station Up-time, and Station Link Up-time (see below)
 - Davis AirLink not writing to the log file during catch-up. Note not all values are available in catch-up
 - Fix leaf wetness graph data when null
+- Fix time format of tags TtempTh, daylength, daylightlength
 
 ### New